Combined dev_params and inst_params, combined expand_params(), apply_inst_params(), resolve_child_deps(),  developed most of child lookup, remove unused functions

+++ flow/plot.py
@@ -462,7 +462,7 @@ def plot_code_density(dout_codes, bins="auto", title="Code Density"):
     )
 
     # Calculate average count and mark deviations
-    avg_count = np.mean(counts)
+    avg_count = float(np.mean(counts))
     ax.axhline(
         avg_count,
         color="r",
